requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.18
django-cors-headers==4.4.0
djangorestframework==3.15.2
ruff==0.9.3
sqlparse==0.5.3
typing-extensions==4.12.2
